+++ src/wasm.h
@@ -928,7 +928,7 @@ public:
   // respected regardless of the value of`hasFeaturesSection`.
   // `hasFeaturesSection` means we read a features section and will emit one
   // too.
-  FeatureSet features = FeatureSet::All;
+  FeatureSet features = FeatureSet::MVP;
   bool hasFeaturesSection = false;
 
   MixedArena allocator;