+++ test/wasm2js/f32.2asm.js
@@ -109,7 +109,7 @@ function asmFunc(global, env, buffer) {
    if (var$2 > Math_fround(.5)) return Math_fround(var$0);
    var$2 = Math_fround(var$1 * Math_fround(.5));
    var$1 = (wasm2js_f32$0 = var$1, wasm2js_f32$1 = var$0, wasm2js_i32$0 = Math_fround(var$2 - Math_fround(Math_floor(var$2))) == Math_fround(0.0), wasm2js_i32$0 ? wasm2js_f32$0 : wasm2js_f32$1);
-  };
+  }
   return Math_fround(var$1);
  }